Ping restructuration : IP or DNS name ?
